JIRA-59: Provide new macros for displaying charts in pages
WIP
Provide a first structure for the work, based for now on the
/rest/gadget/1.0/ endpoint in JIRA. This might change if we see that
this endpoint cannot be used in more recent versions of JIRA.
Right now I focused on piechart: it's expected that quite some work in
there will be abstracted.
Also it's really a PoC right now: there's no tests and almost no doc.
It builds only by skipping lots of checks.